api/requirements.txt
flask~=2.1.1
tensorflow==2.5.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy
Werkzeug~=2.1.1